linted

+++ Handler/ProfileDelete.hs
@@ -18,6 +18,7 @@ module Handler.ProfileDelete where
 
 import Import
 import Handler.Commons
+import Control.Monad (when)
 import qualified Data.Text as T
 import qualified Data.List as L
 import System.Directory
@@ -27,13 +28,13 @@ getProfileDeleteR :: UserId -> Handler Html
 getProfileDeleteR userId = do
   checkRes <- profileCheck userId
   case checkRes of
-    Right user -> do
+    Right user ->
       formLayout $ do
         setTitle "Eidolon :: Delete user profile"
         $(widgetFile "profileDelete")
     Left (errorMsg, route) -> do
       setMessage errorMsg
-      redirect $ route
+      redirect route
 
 postProfileDeleteR :: UserId -> Handler Html
 postProfileDeleteR userId = do
@@ -43,31 +44,29 @@ postProfileDeleteR userId = do
       confirm <- lookupPostParam "confirm"
       case confirm of
         Just "confirm" -> do
-          albumList <- return $ userAlbums user
+          let albumList = userAlbums user
           _ <- mapM (\albumId -> do
             album <- runDB $ getJust albumId
-            case (albumOwner album) == userId of
-              True -> do
-                mediaList <- return $ albumContent album
-                _ <- mapM (\med -> do
-                  commEnts <- runDB $ selectList [CommentOrigin ==. med] []
-                  _ <- mapM (\ent -> runDB $ delete $ entityKey ent) commEnts
-                  medium <- runDB $ getJust med
-                  liftIO $ removeFile (normalise $ L.tail $ mediumPath medium)
-                  liftIO $ removeFile (normalise $ L.tail $ mediumThumb medium)
-                  runDB $ delete med
-                  ) mediaList
-                runDB $ delete albumId
-              False -> return ()
+            when (albumOwner album == userId) $ do
+              let mediaList = albumContent album
+              _ <- mapM (\med -> do
+                commEnts <- runDB $ selectList [CommentOrigin ==. med] []
+                _ <- mapM (runDB . delete . entityKey) commEnts
+                medium <- runDB $ getJust med
+                liftIO $ removeFile (normalise $ L.tail $ mediumPath medium)
+                liftIO $ removeFile (normalise $ L.tail $ mediumThumb medium)
+                runDB $ delete med
+                ) mediaList
+              runDB $ delete albumId
             ) albumList
           runDB $ delete userId
-          liftIO $ removeDirectoryRecursive $ "static" </> "data" </> (T.unpack $ extractKey userId)
+          liftIO $ removeDirectoryRecursive $ "static" </> "data" </> T.unpack (extractKey userId)
           deleteSession "userId"
           setMessage "User deleted successfully"
-          redirect $ HomeR
+          redirect HomeR
         _ -> do
           setMessage "You must confirm the deletion"
           redirect $ ProfileSettingsR userId
     Left (errorMsg, route) -> do
       setMessage errorMsg
-      redirect $ route
+      redirect route
